Start big cleanup by moving methods from IBPartnerDAO to new VATaxIdCheckTargetRepo
They are only needed in de.metas.vatid, so why add them to IBPArtnerDAO in the first place??

TODO
- probably will have to refactor code in VATaxIdCheckTargetRepo even more
  - check what's reall needed to be accessed from outside
  - apply DDD principles
- Address all "// TODO #31060" TODOs
- Migrate possibly existing test-code also into VATaxIdCheckTargetRepo
- Make sure we have proper test-coverage
- Make sure all tests are green
- Revise all classes in de.metas.vatid and consider structuring them in more packages
