+++ backend/de.metas.cucumber/src/test/java/de/metas/cucumber/stepdefs/vatid/VATaxIDCheckProcess_StepDef.java
@@ -44,6 +44,7 @@ import de.metas.security.RoleId;
 import de.metas.user.UserId;
 import de.metas.util.Check;
 import de.metas.util.Services;
+import de.metas.vatid.VATaxIdCheckTargetRepo;
 import de.metas.vatid.process.C_BPartner_VATaxID_Check;
 import io.cucumber.datatable.DataTable;
 import io.cucumber.java.en.Given;
@@ -84,6 +85,8 @@ public class VATaxIDCheckProcess_StepDef
 	@NonNull private final IBPartnerDAO bpartnerDAO = Services.get(IBPartnerDAO.class);
 	@NonNull private final IQueryBL queryBL = Services.get(IQueryBL.class);
 
+	@NonNull private final VATaxIdCheckTargetRepo vaTaxIdCheckTargetRepo;
+	
 	@NonNull private final C_BPartner_StepDefData bpartnerTable;
 
 	/** The PInstance of the last process run, read by {@link #assertPendingCountLogged(int)}. */
@@ -282,7 +285,7 @@ public class VATaxIDCheckProcess_StepDef
 				.addNotInArrayFilter(I_C_BPartner.COLUMNNAME_C_BPartner_ID, keepUnstampedBPartnerIds)
 				.create()
 				.list()
-				.forEach(bpartnerRecord -> bpartnerDAO.stampVATaxIDCheckAttempt(
+				.forEach(bpartnerRecord -> vaTaxIdCheckTargetRepo.stampVATaxIDCheckAttempt(
 						BPartnerId.ofRepoId(bpartnerRecord.getC_BPartner_ID()),
 						attemptedAt));
 
@@ -295,7 +298,7 @@ public class VATaxIDCheckProcess_StepDef
 				.addNotInArrayFilter(I_C_BPartner_Location.COLUMNNAME_C_BPartner_ID, keepUnstampedBPartnerIds)
 				.create()
 				.list()
-				.forEach(bpartnerLocationRecord -> bpartnerDAO.stampVATaxIDCheckAttempt(
+				.forEach(bpartnerLocationRecord -> vaTaxIdCheckTargetRepo.stampVATaxIDCheckAttempt(
 						BPartnerLocationId.ofRepoId(bpartnerLocationRecord.getC_BPartner_ID(), bpartnerLocationRecord.getC_BPartner_Location_ID()),
 						attemptedAt));
 	}
